Several changes
Modified prints made by code during execution, to give a better idea of what the code is doing. Added creation of a text file, called 'modelling_summary' with a summary of modelling at the end of execution. Moved from static allocation of large arrays to dynamic allocation, to allow this code to be executed in MAC OS (this is not yet known to work). Changed names of some variables, so that users can modify with more ease size of different arrays created during execution.